requirements.txt
django
six
django-appconf
beautifulsoup4
